+++ pennylane/tape/__init__.py
@@ -1,0 +1,23 @@
+# Copyright 2018-2020 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""
+This subpackage contains various quantum tapes, which track, queue,
+validate, execute, and differentiate quantum circuits.
+"""
+from .queuing import AnnotatedQueue, Queue, QueuingContext
+from .measure import expval, var, sample, state, probs, MeasurementProcess
+from .operation import mock_operations
+from .circuit_graph import NewCircuitGraph
+from .tapes import QuantumTape, QubitParamShiftTape, CVParamShiftTape, ReversibleTape
+from .qnode import QNode, qnode
