+++ pennylane/tape/tapes/qubit_param_shift.py
